cleaned P5
