--- HW1/P5/title_generator.py
@@ -1,6 +1,0 @@
-import random
-total_nodes = 100000
-
-
-for node in range(1, total_nodes):
-	print "TITLE_" + str(node)
